Add LightCurve.amplitude method
This method returns the peak-to-peak amplitude of a light curve – that
is, the difference between its peak (highest value) and trough (lowest
value). In order to avoid arbitrary results because of outliers caused
by measurement errors, the mean or median of multiple points can be
used as the peak and trough.