Add LEMONdBMiner.amplitudes_by_wavelength method
The first version of the search algorithm does not exclude stars that
have noisy amplitudes, focusing only on whether the amplitudes in all
the filters are directly or inversely correlated with the wavelength.
In other words, we are only interested in whether the star has light
curves in all the photometric filters and their amplitudes increase
or decrease as the wavelength of the photometric filters moves up.

As we intend to make use of a progress bar in the GUI to allow the
user to get a rough estimate of how much time the operation will take,
the algorithm has been implemented as a generator which always returns
something, so that it is possible to know exactly how many items the
generator has stepped through.

A two-element tuple is returned for each star that satisfies the two
conditions, with (a) the ID of the star and (b) a list of two-element
tuples which map each filter to the amplitude of that light curve.
None, on the other hand, is returned if the conditions are not met.