Add LEMONdB.most_similar_magnitude method
In order to determine whether an amplitude is noisy, we will need to
compare it to the standard deviation of the curves of other stars with
similar instrumental magnitudes. Since checking whether a star has a
light curve is not insignificant in terms of performance (an access to
the database is needed in each case and the query is not simple), how
these stars are identified can be implemented as a generator: first
they are sorted by how similar their instrumental magnitude is, and
only then, one by one, we check whether they have a light curve.