Allow LEMONdBMiner.amplitudes_by_wavelength to ignore noisy stars
The method can now, optionally, exclude from the search those stars
with one or more 'noisy' amplitudes. The amplitude A of a star S is
considered noisy if the ratio between A and the median (or mean)
standard deviation of the curves of the N stars in the database with
the brightnesses (instrumental magnitudes) most similar to that of S
is below a threshold.

The last element of the tuple yielded when a star is found becomes a
list of three-element tuples (instead of two-element ones), as they
now must also include the median (or mean) standard deviation of the
comparison standard deviations. In case noisy stars are not ignored,
these standards deviations are set to None.