List actual photometric filters in the labels of the radio buttons
As mentioned in commit fe87f656, the dialog window has "BVI" and "IVB"
as the text of the radio buttons used to select whether the returned
stars must have their amplitudes directly or indirectly, respectively,
correlated with the wavelength. These labels convey perfectly what the
radio buttons do, but it is not hard, and looks much better, to set
them to the real photometric filters in the LEMONdB when the dialog
window (AmplitudesSearchMessageWindow) is created.