Use a progress bar to display the progress of the search
The search for amplitudes correlated with the wavelength, especially
when noisy stars are excluded (the "Filter out those stars with one or
more..." check button) is an expensive task and may take more than the
user would expect. Thus, a gtk.ProgressBar, which grows as stars are
analyzed by LEMONdBMiner.amplitudes_by_wavelength, is used to indicate
the percentage of the work that has been completed.

The search can be canceled at any time by clicking 'Cancel', while
clicking it when no search is in progress closes the window. In order
to avoid running two or more searches in parallel, the 'Ok' button is
disabled as soon as the stars start to be analyzed.