Display search results on a new page of the gtk.Notebook
The AmplitudesSearchPage class receives, one by one, the stars found
by LEMONdBMiner.amplitudes_by_wavelength, populates with them a
gtk.ListStore and creates the gtk.GtkTreeView needed to display its
contents. The get_window method returns a gtk.ScrolledWindow, whose
child widget is the tree view of the stars, that we append as a new
page to the gtk.Notebook of the LEMONJuicerGUI class.