Read value of parameters in message window from configuration file
As we did with the main window, instead of using the default options
(those defined in the Glade XML file) when the GUI starts up, write
them to disk so that they persist across different executions. This
commit adds read-only functionality — which means that the values of
the different parameters are loaded from the configuration file, but
any changes made by the user in the message window are not saved.