Add message window to select the SNR threshold used in plots
The options regarding how light curves are plotted will be from now on
grouped in a new submenu, View → Plots. Move the 'Airmasses' entry to
it and add a new one: 'SNR Threshold'. When clicked, it displays a
message window with a spin button that allows the user to select the
minimum signal-to-noise shown in plots. This commit does not takes the
threshold into account nor stores it in the configuration file, but in
future commits the points of the light curves with a SNR below this
value will not be plotted.