Save and read the SNR threshold to and from the configuration file
The minimum signal-to-noise ratio selected by the user with the spin
button (6274bb3e62) is saved to and read from the configuration file,
so that it persists across different executions. The value is not
updated in the configuration file when 'Cancel' is clicked or the
dialog window destroyed — in other words: only changes in the SNR
threshold that are confirmed with the 'Ok' button are saved.