Update plots when a new SNR threshold is selected
In order to account for the possibility that all the points of a light
curve are below the required minimum signal-to-noise ratio, we now use
a three-row gtk.VBox. The first two are used for the Matplotlib canvas
and navigation toolbar, while the last one contains a gtk.Label with
the error message that is shown when a light curve cannot be plotted
because all its points are below the SNR threshold.

By default, only the first two rows (canvas and toolbar) are visible.
In case the light curve cannot be plotted, the visibility of the rows
in the VBox is reversed and only the error message, "No points have a
signal-to-noise ratio above the current threshold (to modify it, go
to View → Plots → SNR threshold)", is displayed.