Speed up shutdown (quite considerably)
Always use the same handler, LEMONJuicerGUI.handle_quit method, when
the GUI is closed, whether 'Quit' is clicked or the window destroyed.

Take advantage of the fact that the handlers are now unified to speed
up our code: when there are pages in the gtk.Notebook it takes much
longer to close the GUI than when it is empty. We do not know whether
this is how PyGTK works or we are doing something wrong (most probably
the latter), but in any case this can be fixed by removing all the
pages from the notebook before destroying the window.