Use 'button-press-event' signal to switch the photometric filter
In StarDetailsGUI, connecting each photometric filter to the 'clicked'
signal is causing the light curve in the first filter (when sorted by
wavelength) to be plotted first, even if a different one was specified
using the 'init_pfilter' keyword argument (StarDetailsGUI.__init__).
In these cases, the requested filter is plotted, but only after the
first one. Since both plots are made sequentially, this behavior has
gone unnoticed until now.

The reason for this is that the 'clicked' signal is emitted when the
mouse button is pressed and released while the pointer is over the
button. This means that the signal is emitted both when we 'enter'
(press) and 'leave' (unpress) a photometric filter. Since GTK forces
that one of the radio buttons (used to choose the filter) is always
selected, the button of the first filter is always pressed when they
are created (even if the corresponding light curve is not plotted) and
then, when we switch to the filter given in 'init_pfilter', the signal
is emitted twice. The solution is to use the 'button-press-event'
signal, which is only emitted when the mouse button is pressed.