Add explicit method to update the configuration file
We cannot use __del__ to automatically update the configuration file
when the instance is destroyed — this artless approach fails since we
started having more than one reference to the Configuration instance.
We need to use a method that is explicitly called when the GUI shuts
down, as recommended by the wise folks at Stack Overflow:

"You cannot assume that __del__ will ever be called - it is not a
place to hope that resources are automagically deallocated. If you
want to make sure that a (non-memory) resource is released, you should
make a release() or similar method and then call that explicitly [...]"
— http://stackoverflow.com/a/6104568/184363