Fix serious bug in LightCurve.amplitude method
The peak of the light curve is calculated by computing the median (or
arithmetic mean) of the 'npoints' highest magnitudes, but because of a
missing minus sign what we are actually taking are all of them except
for the first 'npoints'. This means that we are basically computing
the median (or mean) of the entire light curve, and therefore the
amplitudes returned by the method are completely wrong.