Add 'Export' button to dump the light curve points to a text file
Include below the list of points of the light curve a gtk.Toolbar with
a single button, 'Export'. When clicked, it displays a message dialog
that allows the user to select which columns of the GtkTreeView (i.e.,
which attributes of the light curve) will be exported to a plain-text
file. The date of observation can be saved in two different formats:
seconds since the Unix epoch or the 24-character representation
returned by the time.asctime function. The path of the file to which
the data is dumped is selected using a gtk.FileChooserDialog, brought
up when 'Save' is clicked.