Make the number of decimal places in exported curves configurable
Use a spin button to allow the user to select how many decimal places
are used when the light curve is exported to a plain-text file. This
spin button has a minimum value of zero (i.e., floats are rounded to
the nearest integer), maximum of eighteen and defaults to eight.