Save and read curve export options to and from the configuration file
Use an event handler for each button in the dialog window (60fe26b4de)
to update the in-memory copy of the configuration file every time the
user changes which attributes of the light curve will be exported to a
plain-text file or the number of decimal places that will be used. In
this manner, the options for how light curves are dumped to text files
persist across different executions.