Assign sequential numbers to amplitudes–wavelength searches
Although, since commit a28012c0e8ca182, AmplitudesSearchPage instances
include a description on the left, they still display the same text in
the tab of the gtk.Notebook: 'Amplitudes ↑' or '↓', depending on how
amplitudes are directly or inversely correlated with the wavelength,
respectively. Use the Roman notation to number the searches made in a
LEMONdB sequentially: the first one is labeled with 'Amplitudes I',
the second, with 'Amplitudes II', and so on.