Always sort columns by wavelength in AmplitudesSearchPage
Matilde Fernández suggests that columns in the AmplitudesSearchPage
instances must always be sorted by the wavelength, so that the user
can easily appreciate how amplitudes increase or decrease with the
wavelength. This approach has the advantages that the photometric
filters are listed in the same order regardless of the type of search,
and that it matches the terms used in the description on the left.