Add AmplitudesSearchPage.toxml method
The method returns the standalone XML representation of the object,
ready to be written to disk. This method will be used to allow the
user to save a search results page to a file so that it can be loaded
back later to resume the work, as suggested by Matilde Fernández.