Remove the parent gtk.GtkBuilder from AmplitudesSearchPage.__init__
Do not require the instantiation method of AmplitudesSearchPage to
receive as its first parameter the gtk.GtkBuilder of the parent GTK
widget. We do not need to share objects between the two widgets, so
the code is cleaner (even if only a tiny bit) if we create our own
instance of gtk.GtkBuilder in the class.