Add 'Save As' button below the description in the AmplitudesSearchPage
Allow the user to export the search results to an XML file, format
which is not only machine-readable (and with which serializing and
de-serializing the AmplitudesSearchPage object is made easier) but
also human-readable. A gtk.FileChooserDialog shows up when the button
is clicked and lets the user browse for the location where the file is
saved. The dialog suggests a name based on the label assigned to the
AmplitudesSearchPage in the LEMONJuicerGUI's gtk.Notebook: lower case,
except for the Roman numerals, and replacing whitespaces with dashes.

Note that, now, the AmplitudesSearchPage.get_label method stores the
last returned title in a class attribute, 'label', so that is can be
used later by the export_to_file method to suggest the filename. This
attribute is not initialized at __init__, so accessing it before the
method is called for the first time will raise the AttributeError
exception. This attribute is needed because the title of the search is
not inherent to the AmplitudesSearchPage, but depends on the number of
searches previously made, track of which is kept by the 'parent'
class, LEMONJuicerGUI.