Allow to open XML files with serialized AmplitudesSearchPage objects
Add a second gtk.FileFilter to the gtk.FileChooserDialog shown by the
LEMONJuicerGUI.handle_open method, so that the user can also select
XML files with serialized AmplitudesSearchPage objects (i.e., with the
results of previous searches for stars whose amplitudes are correlated
with the wavelength). In this first version, all XML files are assumed
to contain a serialized AmplitudesSearchPage object.

Once deserialized, the search result is appended to the gtk.Notebook
and is indistinguishable from other searches. The only difference is
that if the XML file is opened before the LEMONdB the 'row-activated'
signal is ignored — without access to the LEMON database we cannot
show the details of a star when the user double-clicks on it, as we
normally do.