Display both LEMONdBs and XML files in the gtk.FileChooserDialog
Keep the individual filter for each type of file that can be opened by
LEMONJuicerGUI.handle_open (i.e., LEMON databases and XML files with a
serialized AmplitudesSearchPage) but by default use a new filter that
displays both types of files. In this manner, all the files that might
be of interest are listed first, and only if a specific type is needed
the user would need to select one of the other two filters.