Filter XML files by extension, not mimetype
In the gtk.FileChooserDialog, filtering XML files by their mimetype
(application/xml) is not working as expected, as, for example, Glade
files (application/x-glade) are also matched. To avoid this, match XML
files with the .xml extension, but use the gtk.FileFilter.add_custom
method to perform a case-insensitive match.