Insert LEMONdB overview into the first page of the gtk.Notebook
The overview of the LEMONdB must be inserted at the first position of
the gtk.Notebook. Appending it after all the other pages was fine as
long as the LEMON database was guaranteed to be the first page to be
added (and so the -1 and 0 indexes were equivalent), but since commit
123b4ce2d2 it is possible to open a serialized AmplitudesSearchPage
before the LEMONdB.

Therefore, we can no longer append the page with the LEMONdB overview,
as in these cases the page would be inserted at a position other than
zero — and LEMONJuicerGUI.handle_close needs the overview to be in the
first page to be able to detect that the user has closed a LEMONdB and
therefore all the pages in the gtk.Notebook must be removed.