Write the name of the observed objects to the offsets XML file
Add the name of the target astronomical object (read from the FITS
header of each image, most probably from the "OBJECT" or "OBJNAME"
keywords) to the XMLOffset class and store it as an attribute of the
"shifted" element when the offsets are written to disk. This value is
as useful, if not more, than the date of observation and photometric
filter (which are already saved in the XML file) and in order to store
it in the LEMONdB, as we intend to do, we first need to write it in
the offsets XML file.