Store the name of the observed objects in the LEMONdB
Add the name of the target astronomical object as a new column of the
RIMAGE and IMAGES tables (reference and shifted images, respectively)
of the LEMON database. In the case of the shifted images, the name of
the observed object can, since commit b1dc90bc9b846a4, be found in the
offsets XML file, but that of the reference image must still be read
from the FITS header ("OBJECT" or "OBJNAME" keywords, usually).