Store all the data of the shifted image in child XML elements
Although there are no rules about when to use attributes and when to
use child elements, most developers seem to prefer to use attributes
for metadata as opposed to the data itself. Even it that were not the
case, using three attributes ('date', 'filter' and 'object') for the
'shifted' element was resulting in ugly and hard-to-read offsets XML
files. Put this information in child elements — this also requires
that we move the path to the shifted image to its own child element.