Write the FWHM of the shifted images to the offsets XML file
The full width at half maximum of each image is computed by seeing.py
and written to the FITS header. Later, at photometry.py, we compute
the median FWHM of each photometric filter — and thus we have to read
the FWHM of all the images taken with the same filter. However, this
also means that annuli.py accesses the header of all the FITS images
every time a new set of photometric parameters is evaluated, in spite
of the fact that the FWHM of the images is always the same. Save them
to the offsets XML file, so that they are read from the FITS header
only once — when offsets are calculated.