Factor out the reference image in the offsets XML file
Instead of using a 'reference' child for each 'shifted' element, and
since the offsets XML files generated by LEMON use the same reference
for all the images (that is, all the offsets are computed with respect
to the same image), we can factor out the 'reference' element, making
it a child of the root element — in this manner, the reference image,
which is the same for all the images, is listed only once.

Note that this change requires that the XMLOffsetFile class enforces
that all of the XMLOffsets in it contained have the same reference
image. This is achieved by modifying its __init__ method: the class
receives at instantiation the path to the reference image, and
XMLOffsetFile.add checks that the reference image of the XMLOffset
being added is the same. We also need XMLOffsetFile.load, which parses
an XML offsets file and returns an XMLOffsetFile instance.