Save information about the reference image to the offsets XML file
Use the 'image' element not only for each of the shifted images, but
also for the reference image — so that the offsets XML file contains,
in addition to the path to the image, its date of observation, filter,
name of the observed object, FWHM and airmass. In this manner, the XML
file has the same information for all the images, whether reference or
shifted, and photometry.py (and, by extension, annuli.py) does not
need to access the FITS header of the reference image to read it.