Add support for simultaneous observations to the LEMONdB class
This is something that we had been wanting to do for a long time, but
that we have not been forced to implement until now, when trying to
import to a LEMONdB the observations analyzed in 2001 by Carpenter,
Hillenbrand & Skrutskie [1]. These are simultaneous (that is, there
are measures of the stars in multiple photometric filters at exactly
the same time), but the current implementation of the LEMONdB class
forces the Unix time of each image in the database to be unique.

Instead of just the Unix time, use as the primary key of the IMAGES
table both the time and photometric filter. In this manner, there can
be multiple images (and, therefore, photometric measures too) with the
same Unix time, provided that their filters are different. With this,
the LEMONdB class can store simultaneous observations — for example,
data from instruments with a dichroic mirror.

[1] http://adsabs.harvard.edu/abs/2001AJ....121.3160C