Cast Unix time to Python's built-in float in LEMONdB._get_image_id
Force the casting of the first parameter of the query, the date of
observation of the image, to Python's built-in float. This is needed
as SQLite does not support NumPy float, and therefore if it receives
one the "sqlite3.InterfaceError: Error binding parameter - probably
unsupported type" exception will be raised.