Set the tab label to the name of the field when a LEMONdB is opened
Use LEMONdB.field_name to automatically determine the name of the
observed field and set the tab label for the notebook page containing
the list of stars of a LEMONdB to this value. In case the name of the
field cannot be automatically determined (i.e., in the unlikely event
that no prefix is common to the object names of the images), the name
of the .LEMONdB file is used, as we have been doing until now.