Read FWHMs from the XML offsets file, not from the FITS headers
Now that the XML offsets files contain the FWHM of the images, we do
not need to open the FITS images and read it from the header when in
need of computing the median FWHM of each photometric filter. This was
not only unnecessary, but also a bug since the --fwhmk optparse option
was removed in commit a193e8bcdfeb3.