Do not use a list comprehension to discard XMLOffset objects
If we use a list comprehension to take those XMLOffsets whose filter
matches that specified by the user with the --passband option, we end
up with a list, not a XMLOffsetFile, which is what we need. This is a
consequence of commit 884d62a5342f3a0, since which information on the
reference image (such as its airmass or FWHM) is stored as attributes
of the XMLOffsetFile class.