Add METADATA table
Although we are using SQLite, we do not need a relational model to
store information about the LEMONdB such as who run the pipeline, on
which computer this was done or the date of creation of the database.
This records (key-value pairs) will be stored in this table, where we
do not allow, by means of the UNIQUE constraint, more than one row for
each different record — i.e., there cannot be multiple rows with the
same value in the 'key' column.