Add LEMONdB.id property — unique identifier of the LEMONdB
This property encapsulates the LEMONdB._get_id and _set_id non-public
methods, getter and setter of the 'ID' record in the METADATA table.
It is intended to store a unique identifier of the LEMONdB, which we
set after doing photometry: the MD5 hash (a 32-digit hexadecimal
number) of the concatenation, in this order, of the LEMONdB.date,
author and hostname properties.