Write identifier of the LEMONdB to the AmplitudesSearchPage XML file
Add a second attribute to the root element of the XML representation
of the AmplitudesSearchPage object: 'database_id', which is assigned
the value of LEMONdB.id, the unique identifier of the database. In
this manner, the results of the search can be linked to a specific
database — this will allow us to prevent the user from loading an XML
file with results for a LEMONdB other than the one currently open.