Do not load AmplitudesSearchPage XML files with a different identifier
Take into account the unique identifier of the LEMONdB, stored in the
'database_id' attribute of the root element, when an XML file with a
serialized AmplitudesSearchPage object is opened. If the identifier
does not match that of the current LEMONdB, if any, an error message
is displayed, informing the user that the XML file cannot be opened
because it contains results for a different database.