Use the observed field to suggest a filename in ExportCurveDialog
Include the name of the observed field in the filename suggested by
the gtk.FileChooserDialog when the light curve of a star is saved to a
plain-text file in the ExportCurveDialog.run method. This requires the
modification of __init__, as it must receive an additional argument
with the name of the field, determined by LEMONdB.field_name.