Use the observed field to suggest a filename in AmplitudesSearchPage
Take the name of the observed field into consideration when suggesting
a filename in AmplitudesSearchPage.export_to_file. This requires the
modification of __init__, to receive an additional argument with the
name of the field, as well as the addition of a new element, 'field',
to the structure of the XML to which AmplitudesSearchPage objects are
serialized.