Add CANDIDATE_PARAMETERS table
This table will be used to store in the LEMONdB the contents of the
XML file, output by the annuli module, with the candidate parameters
for aperture photometry that were evaluated for each photometric
filter. It maps each set of photometric parameters (stored in the
PHOTOMETRIC_PARAMETERS table) to a filter (PHOTOMETRIC_FILTERS) to the
standard deviation of the light curves of the constant stars that were
used to determine how good the parameter set is.