Add LEMONdB.add_candidate_pparams method
The method adds a new record to the CANDIDATE_PARAMETERS table. Note
that a set of photometric parameters (aperture, annulus and dannulus)
can only map to a standard deviation for each photometric filter, so
adding two CandidateAnnuli objects for the same filter with the same
aperture, annulus and dannulus, but different standard deviations,
will replace the first one with the second.