Add LEMONdB.get_candidate_pparams method
The method returns a list with all the CandidateAnnuli objects that
have been stored in the CANDIDATE_PARAMETERS table for a specific
photometric filter. These are sorted in increasing order by the
standard deviation of the light curves of the constant stars that were
used to determine how good the parameters are (CandidateAnnuli.stdev).