Remove 'best_only' parameter from CandidateAnnuli.xml_load method
Now that we intend to store all the candidate photometric parameters
in the LEMONdB, the 'best_only' parameter of CandidateAnnuli.xml_load
is not necessary anymore, as we will always need to parse the entire
XML file. The small size of these XML files means that performance
will not suffer, neither now nor in the foreseeable future, even if
all the CandidateAnnuli, and not only the best ones, are loaded.