Sort lists returned by CandidateAnnuli.xml_load by standard deviation
The CandidateAnnuli.xml_load static method returns a dictionary which
maps each photometric filter to a list of CandidateAnnuli objects. As,
since commit d8710928a58c2f2459bf, the method always returns all the
candidate photometric parameters in the XML file, we need a way to
straightforwardly identify which one is the best (and therefore the
one that we should use to do photometry) for each filter.

To achieve this, sort these lists in increasing order by the standard
deviation of the light curves of the constant stars that were used to
determine how good the parameters are (CandidateAnnuli.stdev). In this
manner, the first CandidateAnnuli of each list is always the best one.