Fix bug in ImageTest.nrandom class method
The invariant that in no case two returned random Images have the same
Unix time and photometric filter does not always hold: this is exactly
what happens if the randomly-selected Unix time to reuse exhausts all
the photometric filters — in other words, if the Unix time has already
been reused for all possible filters. To fix this it is necessary not
to modify the Unix time of the random Image until we have made sure
that there is a photometric filter in which it has not yet been used.