Improve SQL query in LEMONdB._star_pfilters and pfilters methods
The current implementation of the inner join (PHOTOMETRY, IMAGES and
PHOTOMETRIC_FILTERS tables) extracts the wavelengths of all of the
rows in the first table, PHOTOMETRY, and then filters out duplicates
with the DISTINCT clause. However, since what we need to identify are
nothing more than the rows in IMAGES for which there is a matching row
in PHOTOMETRY, the number of records that are combined can be reduced
considerably if we use a subquery to extract the distinct IDs of the
images in PHOTOMETRY before proceeding with the first inner join.