Add LEMONdB.analyze method
The ANALYZE command gathers statistics about tables and indexes and
stores the collected information in internal tables of the database
where the query optimizer can access and use it to help make better
query planning choices. The current transaction is committed and a
new one automatically started when the commands finishes.

The statistics gathered by ANALYZE are not automatically updated as
the content of the database changes. If the content of the LEMONdB
changes significantly, we should consider rerunning the command in
order to update the statistics.

SQL As Understood By SQLite: ANALYZE
https://www.sqlite.org/lang_analyze.html