Remove redundant indexes
The two-column index IMAGES(wavelength, unix_time) makes unnecessary
the existence of the other one, IMAGES(wavelength): the left-most
column is the primary key used for ordering the rows in the index, and
the right-most used to break ties, if any. There is no need of keeping
the single-column index when the two-colum one allows for identical
functionality. The same applies to the PERIODS(star_id, wavelength)
and PERIODS(star_id) indexes — we can get rid of the latter.