Replace phot_by_star with a two-column index (star_id, image_id)
A new index, PHOTOMETRY(star_id, image_id), is needed to speed up the
query in the LEMONdB.get_photometry method. This addition means that
PHOTOMETRY(star_id) can be removed: they both start with the same
column, so queries that would use the single-column index can also use
the two-column one — it is not necessary for every column of an index
to appear in a WHERE clause term in order for that index to be used.