Add INDEXED BY clauses to the query in LEMONdB.get_photometry
The INDEXED BY clause specifies the index that must be used in order
to look up values on the preceding table. If the index does not exist
or cannot be used for the query, then the preparation of the statement
fails. Add it to the query in LEMONdB.get_photometry to guarantee that
the phot_by_star_image and img_by_wavelength_time indexes are used.

This clause, an SQL extension found only in SQLite, is not intended
for use in tuning the performance of a SQL query, but to raise a
run-time error if a schema change, such as dropping or creating an
index, causes the query plan for a time-sensitive query to change.

SQL As Understood By SQLite: INDEXED BY
https://www.sqlite.org/lang_indexedby.html