Replace curve_by_star with a two-column index (star_id, image_id)
The LIGHT_CURVES(star_id, image_id) index is necessary to speed up the
first of the two SQL queries in LEMONdB.get_light_curve. As was done
in commit 6a3150d2f6f4d, this means that index LIGHT_CURVES(star_id),
which starts with the same row, becomes redundant and can be removed.