Add range checking to Catalog.flag_saturated class method
Internal flags, accessible through the FLAGS catalog parameter, are,
although coded in decimal, computed by SExtractor as the sum of the
first eight powers of two (exponents 1-7). Therefore, the minimum
possible value is zero (0b00000000) and the maximum 255 (0b11111111).
Raise the ValueError exception in case the decimal value of the flag
is outside of this range.