Fix bug in Catalog.flag_saturated
Although Catalog.flag_saturated returns the correct value for all
possible SExtractor flags, the algorithm does not work correctly for
flags smaller than four: the length of the string returned by bin() is
always greater than or equal to three, because of the '0b' prefix, and
therefore the value checked is not that of the third least significant
bit, but the '0' or the 'b' in the prefix. For example, bin(2) returns
'0b10', and '0b10'[-3] is 'b', which is not equal to '1', so False is
returned — although only by chance.

In order to fix this, get rid of the '0b' prefix and fill in with
zeros so that SExtractor flags are always represented as eight-bit
binary numbers.