Make the Pixel class immutable
The implementation of the Pixel class can be enormously simplified if,
instead of 'object', we use a named tuple (collections.namedtuple) as
its base class. In this manner, Pixel becomes a tuple-like class whose
fields (the x- and y-coordinates) are accessible by attribute lookup.
Note that this means that Pixel becomes immutable — it is now a tuple
subclass with named fields. We will need to use _replace() to get a
new instance where one or the two coordinates are modified.