Read SExtractor catalog only once
Avoid loading the entire SExtractor catalog into memory every time
that the index of a parameter must be determined. Instead, do it only
once, at instantiation, and make Catalog._find_column take as inputs
not only the name of the parameter but also the contents of the file.
As a consequence of this change, the Catalog._read_catalog method is
no longer needed and can be removed.