Make Catalog._find_column static
The method does not need access to instance data, as the contents of
the SExtractor catalog are now taken as input, so simplify the code by
making it static. Note that this means that we can no longer cache the
indexes internally — this is not much of a problem, nevertheless, as
we plan to be able to use our memoization decorator soon.