Find the index of each parameter only once in Catalog._load_stars
The column in which each parameter is located in a catalog is always
the same, so we do not need to call Catalog._find_column() for each
line (that is, detected source) in the file. Do it just once, at the
beginning of Catalog._load_stars, before iterating through the
non-commented lines of the SExtractor catalog.