Add Catalog.from_sequence class method
This class method returns a Catalog that contains the Star objects
passed as arguments. We will need to use it because, now that Catalog
uses tuple as its base class, neither item assignment nor deletion are
possible, so we cannot modify a SExtractor catalog after it is loaded
into memory. In case we need to discard some stars (e.g., those above
the saturation threshold) this method can be used to create another
Catalog with the non-discarded ones — note that we could use a list
comprehension, but that would give us a list, not a Catalog object.