Use Catalog.from_sequence to filter out stars too close to the edges
Now that it subclasses tuple, it is no longer possible to iterate a
Catalog backwards removing those stars that are too close to the image
edges. Therefore, we first have to identify the stars that are not to
be ignored (i.e., those that are not too close to a border) and pass
them to Catalog.from_sequence() to create a new Catalog object
containing only them.