Add ignore_del_exceptions context manager
Due to the precarious circumstances under which __del__() methods are
invoked, uncaught exceptions that occur during their execution are
ignored, and a warning is printed to standard error instead. This
context manager makes it possible to prevent these (and only these)
warning messages from being sent to standard error without having to
modify the code that raises the uncaught exception.