Add high-level command 'lemon'
In the same way that Git differentiates between low-level ('plumbing')
commands, designed to be chained together Unix style, and high-level
('porcelain') ones, which are more user-friendly, LEMON will provide a
single point of entry to all of the functionality implemented by the
pipeline: the 'lemon' command (note the absence of '.py' extension).

The task to be performed will be specified by a subcommand (e.g.,
"lemon photometry"). The corresponding module (such as photometry.py)
will be then imported and its main() method called, with the rest of
arguments passed to it.

In case no arguments are provided, this first version of the 'lemon'
command shows a help message (also disturbingly similar to that of
Git) with all the available subcommands.