Add --help option
Display the show message not only if no arguments are passed, but also
when the --help option is used. Note that we are detecting this option
without optparse, as that approach would not work if other options
(for example, --pixels when doing photometry) were used — these are
recognized by the option parsers defined in each module (such as
photometry.py), but they are unknown to this hypothetical upper-level
parser, which would generate an error message along the lines of
"error: no such option: --pixels" and exit with error status 2.