Show error message and exit if subcommand is unrecognized
The name of the LEMON subcommand (task) to execute must exactly match,
case included, one of the following: import, seeing, offsets, mosaic,
astrometry, annuli, photometry, diffphot and periods. If that is not
the case, an error message is printed to the standard error and the
program exits with a code of 1. Note that the subcommands match the
name of the modules that implement the different pipeline stages.