Show suggestions, if any, when the subcommand does not exist
When the subcommand specified by the user is unrecognized, compare it
to the list of available tasks and compute the list of the best 'good
enough' matches using the difflib.get_close_matches function. These
are displayed as suggestions to the user — who might have, for
example, typed 'pohtometry' instead of 'photometry'.