Import the task module and execute its main() function
If the subcommand is a valid one, import the LEMON module of the same
name (e.g., photometry.py when the photometry subcommand is specified)
and run its main() function. Note that we need to use the __import__
built-in function, instead of the 'import' statement, as the name of
the module is only known at runtime.