Append the name of the subcommand to sys.argv[0]
Replace sys.argv[0] with the name of the program (lemon) followed by
the specified subcommand (for example, diffphot). Although sys.argv[0]
is not passed to the main() function of the imported module, that is
where optparse looks up the name of the script when the brief summary
of options is displayed (that is, when -h or --help are used).

In short: make the optpase message start with something like, for
example, "Usage: lemon diffphot [OPTION]... INPUT_DB", instead of
"Usage: lemon [OPTION]... INPUT_DB" — note how the subcommand is
included in the former case.