Catch UnboundLocalError raised in FITSImage.prefix
FITSImage.prefix raises UnboundLocalError when the basename of the
image, once the prefix is stripped, is one character long or an empty
string. When this happens, the for loop does not execute but the
'index' variable, although undefined, is still accessed later on.
Account for this possibility, returning an empty string.