Fix call to astromatic.sextractor
The invocation of the astromatic.sextractor method is incorrect, as
it does not return a two-element tuple anymore, but only the path to
the SExtractor catalog. The astromatic.SExtractorError exception is
raised if the exit status is other than zero, so there is no need to
return this value along with the path to the catalog.